backend/requirements.txt
Django==5.0.6
django-ninja==1.3.0
django-cors-headers==4.3.1
django-eventstream==5.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
openai==1.35.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-decouple==3.8
jinja2==3.1.4
whitenoise==6.6.0
gunicorn==22.0.0
pyyaml==6.0.1
faker==25.2.0
pytest==8.2.1
pytest-django==4.8.0
pytest-asyncio==0.23.7
httpx==0.27.0
